// app/frontend/src/shared/schemas/matchesSchemas.ts
export enum MatchStatus {
    PENDING = 'pending',
    IN_PROGRESS = 'in_progress',
    FINISHED = 'finished'
}
export const MatchStatusSchema = z.nativeEnum(MatchStatus);
// ...
export const MatchBaseSchema = z.object({
    id: z.number().int(),
    matchId: z.string(), // ou i min
    player1_id: z.number().int(),
    player2_id: z.number().int(),
    player1_socket: z.string().nullable(),
    player2_socket: z.string().nullable(),
    player1_score: z.number().int().default(0),
    player2_score: z.number().int().default(0),
    winner_id: z.number().nullable(),
    win_type: z.string().nullable(),
    created_at: z.string(),
    tournament_id: z.string().nullable(),
    status: MatchStatusSchema.default(MatchStatus.PENDING)
});
export type Match = z.infer<typeof MatchBaseSchema>;
// ...
export const MatchUserIdParamsSchema = z.object({
    userId: z.string().regex(/^\d+$/, "User ID must be a positive integer."),
});

export type MatchUserIdParams = z.infer<typeof MatchUserIdParamsSchema>;

export const GetMatchByUserIdRouteSchema = {
    params: MatchUserIdParamsSchema,
    response: {
        200: z.array(MatchBaseSchema),
        400: z.object({ error: z.string() }),
        500: z.object({ error: z.string() })
    }
}